Handler function for /register
Code seemed to get too long so I moved to a separate file.

Added use of SQL functions to check for a user, as well as to add user to DB.

Added bcrypt password hashing.

'exists' is type user, but since I wanted to use the user variable later on I gave 'exists' a more meaningful name for what it does.