isAuthentcated not working
isAuthenticated doesn't seem to work when trying to access a restricted page without logging in.
It returns true even when it shouldn't. Could be because of typing issues.

Changed ensureAuthenticated to simply check if req.user is defined, seems to fix the problem.

Also only returns next() after successful authentication.